+++ strategy/bear.go
@@ -1,4 +1,4 @@
-package main
+package strategy
 
 import (
 	"context"
@@ -14,11 +14,14 @@ import (
 
 	binance_connector "github.com/binance/binance-connector-go"
 	"github.com/wferreirauy/binance-bot/ai"
+	"github.com/wferreirauy/binance-bot/exchange"
+	"github.com/wferreirauy/binance-bot/indicator"
 	"github.com/wferreirauy/binance-bot/config"
 	"github.com/wferreirauy/binance-bot/tui"
 )
 
-func BullTrade(
+// BearTrade implements a sell-high-buy-low strategy for bearish markets.
+func BearTrade(
 	configFile string,
 	symbol string,
 	qty float64,
@@ -37,8 +40,8 @@ func BullTrade(
 	if err != nil {
 		log.Fatal(err)
 	}
-	period := cfg.HistoricalPrices.Period     // length period for moving average
-	interval := cfg.HistoricalPrices.Interval // time intervals of historical prices for trading
+	period := cfg.HistoricalPrices.Period
+	interval := cfg.HistoricalPrices.Interval
 
 	// refresh interval for price polling (default 10 seconds)
 	refreshSecs := cfg.RefreshInterval
@@ -48,9 +51,9 @@ func BullTrade(
 	refreshInterval := time.Duration(refreshSecs) * time.Second
 
 	// initialize binance api client
-	client := binance_connector.NewClient(apikey, secretkey, baseurl)
+	client := binance_connector.NewClient(exchange.APIKey, exchange.SecretKey, exchange.BaseURL)
 
-	// validate symbol in format 0-9A-Z/0-9A-Z
+	// validate symbol
 	if re := regexp.MustCompile(`(?m)^[0-9A-Z]{1,8}/[0-9A-Z]{2,8}$`); !re.Match([]byte(symbol)) {
 		log.Fatal("error parsing ticker: must match ^[0-9A-Z]{1,8}/[0-9A-Z]{2,8}$")
 	}
@@ -77,7 +80,7 @@ func BullTrade(
 	}
 
 	// initialize TUI dashboard
-	dash := tui.NewDashboard("BULL", symbol)
+	dash := tui.NewDashboard("BEAR", symbol)
 
 	// initialize file logger
 	fl, err := tui.NewFileLogger("binance-bot.log")
@@ -100,7 +103,7 @@ func BullTrade(
 		if aiOrch != nil {
 			dash.LogInfo("AI Agents: [green]ENABLED[-]")
 		}
-		bullTradeLoop(dash, client, cfg, aiOrch, symbol, ticker, scoin, dcoin, qty, stopLoss, takeProfit, buyFactor, sellFactor, roundPrice, roundAmount, max_ops, period, interval, refreshInterval)
+		bearTradeLoop(dash, client, cfg, aiOrch, symbol, ticker, scoin, dcoin, qty, stopLoss, takeProfit, buyFactor, sellFactor, roundPrice, roundAmount, max_ops, period, interval, refreshInterval)
 	}()
 
 	if err := dash.Run(); err != nil {
@@ -108,7 +111,7 @@ func BullTrade(
 	}
 }
 
-func bullTradeLoop(
+func bearTradeLoop(
 	dash *tui.Dashboard,
 	client *binance_connector.Client,
 	cfg *config.Config,
@@ -120,18 +123,18 @@ func bullTradeLoop(
 	interval string,
 	refreshInterval time.Duration,
 ) {
-	var buyPrice float64
+	var sellPrice float64
 	var operation = 1
 	var consecutiveSL int // tracks consecutive stop-loss exits for cooldown
 
 	for range max_ops {
 		dash.SetOperation(operation)
-		qty = roundFloat(qty, roundAmount)
+		qty = indicator.RoundFloat(qty, roundAmount)
 
-		//// buy ////
-		dash.SetPhase("SCANNING BUY")
+		//// sell (bear entry) ////
+		dash.SetPhase("SCANNING SELL")
 		for {
-			ohlcv, err := getHistoricalOHLCV(client, ticker, interval, period)
+			ohlcv, err := exchange.GetHistoricalOHLCV(client, ticker, interval, period)
 			if err != nil {
 				dash.LogError(fmt.Sprintf("OHLCV fetch: %v", err))
 				time.Sleep(refreshInterval)
@@ -143,7 +146,7 @@ func bullTradeLoop(
 			dash.UpdatePrice(price, prevPrice, roundPrice)
 
 			// tendency
-			tendency, err := getTendency(client, ticker, cfg.Tendency.Interval, period)
+			tendency, err := exchange.GetTendency(client, ticker, cfg.Tendency.Interval, period)
 			if err != nil {
 				dash.LogError(fmt.Sprintf("Tendency: %v", err))
 				time.Sleep(refreshInterval)
@@ -151,11 +154,11 @@ func bullTradeLoop(
 			}
 
 			// indicators
-			dema := calculateDEMA(ohlcv.Closes, cfg.Indicators.Dema.Length)
+			dema := indicator.CalculateDEMA(ohlcv.Closes, cfg.Indicators.Dema.Length)
 			currentDema := dema[len(dema)-1]
-			rsi := calculateRSI(ohlcv.Closes, cfg.Indicators.Rsi.Length)
-			macdLine, signalLine := calculateMACD(ohlcv.Closes, cfg.Indicators.Macd.FastLength, cfg.Indicators.Macd.SlowLength, cfg.Indicators.Macd.SignalLength)
-			bb, err := CalculateBollingerBands(ohlcv.Closes, cfg.Indicators.BollingerBands.Length, cfg.Indicators.BollingerBands.Multiplier)
+			rsi := indicator.CalculateRSI(ohlcv.Closes, cfg.Indicators.Rsi.Length)
+			macdLine, signalLine := indicator.CalculateMACD(ohlcv.Closes, cfg.Indicators.Macd.FastLength, cfg.Indicators.Macd.SlowLength, cfg.Indicators.Macd.SignalLength)
+			bb, err := indicator.CalculateBollingerBands(ohlcv.Closes, cfg.Indicators.BollingerBands.Length, cfg.Indicators.BollingerBands.Multiplier)
 			if err != nil {
 				dash.LogError(fmt.Sprintf("BollingerBands: %v", err))
 			}
@@ -164,17 +167,17 @@ func bullTradeLoop(
 			distanceToUpper := math.Abs(currentDema - upperBand)
 			distanceToLower := math.Abs(currentDema - lowerBand)
 
-			// MACD cross description
-			macdCross := "BEARISH"
-			if macdLine[len(macdLine)-2] <= signalLine[len(signalLine)-2] && macdLine[len(macdLine)-1] > signalLine[len(signalLine)-1] {
-				macdCross = "BULLISH"
+			// MACD cross
+			macdCross := "BULLISH"
+			if macdLine[len(macdLine)-2] >= signalLine[len(signalLine)-2] && macdLine[len(macdLine)-1] < signalLine[len(signalLine)-1] {
+				macdCross = "BEARISH"
 			}
 
 			// ADX
 			var adxVal float64
 			var adxStrong bool
 			if cfg.Indicators.Adx.Period > 0 {
-				adx := calculateADX(ohlcv.Highs, ohlcv.Lows, ohlcv.Closes, cfg.Indicators.Adx.Period)
+				adx := indicator.CalculateADX(ohlcv.Highs, ohlcv.Lows, ohlcv.Closes, cfg.Indicators.Adx.Period)
 				if len(adx) > 0 {
 					adxVal = adx[len(adx)-1]
 				}
@@ -187,7 +190,7 @@ func bullTradeLoop(
 			var currentVolume, avgVolume float64
 			var volumeConfirmed bool
 			if cfg.Indicators.Volume.MaPeriod > 0 {
-				volumeMA := calculateSMA(ohlcv.Volumes, cfg.Indicators.Volume.MaPeriod)
+				volumeMA := indicator.CalculateSMA(ohlcv.Volumes, cfg.Indicators.Volume.MaPeriod)
 				currentVolume = ohlcv.Volumes[len(ohlcv.Volumes)-1]
 				if len(volumeMA) > 0 {
 					avgVolume = volumeMA[len(volumeMA)-1]
@@ -217,43 +220,43 @@ func bullTradeLoop(
 					ADX: adxVal, Volume: currentVolume, AvgVolume: avgVolume,
 				}
 				ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-				consensus, err := aiOrch.Analyze(ctx, snapshot, "BULL")
+				consensus, err := aiOrch.Analyze(ctx, snapshot, "BEAR")
 				cancel()
 				if err != nil {
 					dash.LogError(fmt.Sprintf("AI: %v", err))
 				} else {
 					updateDashAI(dash, consensus)
-					aiApproved = consensus.ShouldBuy() || consensus.FinalSignal == ai.SignalHold
+					aiApproved = consensus.ShouldSell() || consensus.FinalSignal == ai.SignalHold
 				}
 			}
 
-			// Higher-timeframe trend gate: block BULL entry if HTF trend is not up
+			// Higher-timeframe trend gate: block BEAR entry if HTF trend is not down
 			if cfg.Tendency.HTFEnabled && cfg.Tendency.HTFInterval != "" {
-				htfTendency, htfErr := getTendency(client, ticker, cfg.Tendency.HTFInterval, period)
+				htfTendency, htfErr := exchange.GetTendency(client, ticker, cfg.Tendency.HTFInterval, period)
 				if htfErr != nil {
 					dash.LogError(fmt.Sprintf("HTF Tendency: %v", htfErr))
-				} else if htfTendency != cfg.Tendency.Direction {
-					dash.LogInfo(fmt.Sprintf("[red]HTF GATE[-] %s trend is [red]%s[-] on %s — skipping BULL entry",
+				} else if htfTendency != "down" {
+					dash.LogInfo(fmt.Sprintf("[red]HTF GATE[-] %s trend is [green]%s[-] on %s — skipping BEAR entry",
 						symbol, htfTendency, cfg.Tendency.HTFInterval))
 					time.Sleep(refreshInterval)
 					continue
 				}
 			}
 
-			// when to buy — scalp mode uses scoring; classic mode requires all conditions
-			var shouldBuy bool
+			// when to sell (bear entry) — scalp mode uses scoring; classic mode requires all conditions
+			var shouldSell bool
 			if cfg.ScalpMode.Enabled {
 				score := 0
-				if rsi[len(rsi)-1] < float64(cfg.Indicators.Rsi.UpperLimit) {
+				if rsi[len(rsi)-1] > float64(cfg.Indicators.Rsi.LowerLimit) {
 					score++
 				}
-				if macdLine[len(macdLine)-1] > signalLine[len(signalLine)-1] {
-					score++ // MACD above signal (not requiring exact crossover)
+				if macdLine[len(macdLine)-1] < signalLine[len(signalLine)-1] {
+					score++ // MACD below signal (not requiring exact crossover)
 				}
-				if tendency == cfg.Tendency.Direction {
+				if tendency == "down" {
 					score++
 				}
-				if distanceToLower < distanceToUpper {
+				if distanceToUpper < distanceToLower {
 					score++
 				}
 				if adxStrong {
@@ -266,71 +269,64 @@ func bullTradeLoop(
 				if minScore <= 0 {
 					minScore = 3
 				}
-				shouldBuy = score >= minScore && aiApproved
-				if shouldBuy {
+				shouldSell = score >= minScore && aiApproved
+				if shouldSell {
 					dash.LogInfo(fmt.Sprintf("[yellow]Scalp entry: score %d/%d (min %d)[-]", score, 6, minScore))
 				}
 			} else {
-				shouldBuy = rsi[len(rsi)-1] < float64(cfg.Indicators.Rsi.UpperLimit) &&
-					macdLine[len(macdLine)-2] <= signalLine[len(signalLine)-2] &&
-					macdLine[len(macdLine)-1] > signalLine[len(signalLine)-1] &&
-					tendency == cfg.Tendency.Direction &&
-					distanceToLower < distanceToUpper &&
+				shouldSell = rsi[len(rsi)-1] > float64(cfg.Indicators.Rsi.LowerLimit) &&
+					macdLine[len(macdLine)-2] >= signalLine[len(signalLine)-2] &&
+					macdLine[len(macdLine)-1] < signalLine[len(signalLine)-1] &&
+					tendency == "down" &&
+					distanceToUpper < distanceToLower &&
 					adxStrong && volumeConfirmed && aiApproved
 			}
-			if shouldBuy {
+			if shouldSell {
 
-				dash.SetPhase("BUYING")
-				buy, err := TradeBuy(symbol, qty, price, buyFactor, roundPrice)
+				dash.SetPhase("SELLING")
+				sell, err := TradeSell(symbol, qty, price, sellFactor, roundPrice)
 				if err != nil {
-					dash.LogError(fmt.Sprintf("BUY order failed: %v", err))
+					dash.LogError(fmt.Sprintf("SELL order failed: %v", err))
 					return
 				}
-				buyOrder := reflect.ValueOf(buy).Elem()
-				orderId := buyOrder.FieldByName("OrderId").Int()
-				orderPrice := buyOrder.FieldByName("Price").String()
-				buyPrice, _ = strconv.ParseFloat(orderPrice, 64)
+				sellOrder := reflect.ValueOf(sell).Elem()
+				orderId := sellOrder.FieldByName("OrderId").Int()
+				orderPrice := sellOrder.FieldByName("Price").String()
+				sellPrice, _ = strconv.ParseFloat(orderPrice, 64)
 
-				dash.LogOrder(fmt.Sprintf("[green::b]BUY[-] %f %s @ [white::b]%.*f[-] %s = %.*f %s",
-					qty, scoin, roundPrice, buyPrice, dcoin, roundPrice, buyPrice*qty, dcoin))
+				dash.LogOrder(fmt.Sprintf("[red::b]SELL[-] %f %s @ [white::b]%.*f[-] %s = %.*f %s",
+					qty, scoin, roundPrice, sellPrice, dcoin, roundPrice, sellPrice*qty, dcoin))
 
-				if getor, err := GetOrder(ticker, orderId); err == nil {
-					dash.LogInfo(fmt.Sprintf("BUY order #%d - Status: %s", getor.OrderId, getor.Status))
+				if getor, err := exchange.GetOrder(ticker, orderId); err == nil {
+					dash.LogInfo(fmt.Sprintf("SELL order #%d - Status: %s", getor.OrderId, getor.Status))
 				}
 
-				for {
-					if getor, err := GetOrder(ticker, orderId); err == nil {
-						if getor.Status == "FILLED" {
-							dash.LogOrder("[green::b]BUY order filled![-]")
-							break
-						}
-					}
-					time.Sleep(refreshInterval)
-				}
+				waitOrderFilled(dash, ticker, orderId, "[red::b]SELL[-] order filled!", refreshInterval)
 				break
 			}
 			time.Sleep(refreshInterval)
 		}
 
-		postBuyDelay := 30
+		postSellDelay := 30
 		if cfg.ScalpMode.Enabled && cfg.ScalpMode.PostBuyDelay > 0 {
-			postBuyDelay = cfg.ScalpMode.PostBuyDelay
+			postSellDelay = cfg.ScalpMode.PostBuyDelay
 		}
-		time.Sleep(time.Duration(postBuyDelay) * time.Second)
+		time.Sleep(time.Duration(postSellDelay) * time.Second)
 
-		//// sell ////
-		dash.SetPhase("MONITORING SELL")
-		highestPrice := buyPrice
+		//// buy back (bear exit) ////
+		dash.SetPhase("MONITORING BUY-BACK")
+		lowestPrice := sellPrice
+		sellProceeds := sellPrice * qty
 		exitType := "" // tracks how position was closed: "tp", "ts", "sl"
 
 		for {
-			ohlcv, err := getHistoricalOHLCV(client, ticker, interval, period)
+			ohlcv, err := exchange.GetHistoricalOHLCV(client, ticker, interval, period)
 			if err != nil {
 				dash.LogError(fmt.Sprintf("OHLCV fetch: %v", err))
 				time.Sleep(refreshInterval)
 				continue
 			}
-			rsiprices, err := getHistoricalPrices(client, ticker, cfg.Indicators.Rsi.Interval, period)
+			rsiprices, err := exchange.GetHistoricalPrices(client, ticker, cfg.Indicators.Rsi.Interval, period)
 			if err != nil {
 				dash.LogError(fmt.Sprintf("RSI prices: %v", err))
 				time.Sleep(refreshInterval)
@@ -339,37 +335,38 @@ func bullTradeLoop(
 
 			price := ohlcv.Closes[len(ohlcv.Closes)-1]
 			prevPrice := ohlcv.Closes[len(ohlcv.Closes)-2]
-			rsi := calculateRSI(rsiprices, cfg.Indicators.Rsi.Length)
+			rsi := indicator.CalculateRSI(rsiprices, cfg.Indicators.Rsi.Length)
 			dash.UpdatePrice(price, prevPrice, roundPrice)
 
-			// update indicators panel with sell-phase data
-			pnl := (price - buyPrice) / buyPrice * 100
+			// update indicators with P&L
+			pnl := (sellPrice - price) / sellPrice * 100
 			dash.UpdateIndicators(&tui.IndicatorData{
 				RSI: rsi[len(rsi)-1], RSIUpperLimit: cfg.Indicators.Rsi.UpperLimit, RSILowerLimit: cfg.Indicators.Rsi.LowerLimit,
 				Tendency: fmt.Sprintf("P&L: %+.2f%%", pnl),
 			})
 
-			if price > highestPrice {
-				highestPrice = price
+			if price < lowestPrice {
+				lowestPrice = price
 			}
 
-			// trailing stop-loss
+			// trailing stop (inverse for bear)
 			if cfg.TrailingStop.Enabled {
-				activationPrice := buyPrice * (1 + cfg.TrailingStop.ActivationPct/100)
-				if highestPrice >= activationPrice {
-					trailingStopPrice := highestPrice * (1 - cfg.TrailingStop.TrailingPct/100)
-					if price <= trailingStopPrice {
+				activationPrice := sellPrice * (1 - cfg.TrailingStop.ActivationPct/100)
+				if lowestPrice <= activationPrice {
+					trailingStopPrice := lowestPrice * (1 + cfg.TrailingStop.TrailingPct/100)
+					if price >= trailingStopPrice {
 						dash.SetPhase("TRAILING STOP")
-						sell, err := TradeMarketSell(symbol, roundFloat(qty*0.998, roundAmount), price, roundPrice)
+						buyBackQty := indicator.RoundFloat(sellProceeds/price, roundAmount)
+						buy, err := TradeMarketBuy(symbol, buyBackQty, price, roundPrice)
 						if err != nil {
-							dash.LogError(fmt.Sprintf("Trailing-Stop MARKET SELL failed: %v", err))
+							dash.LogError(fmt.Sprintf("Trailing-Stop MARKET BUY failed: %v", err))
 							return
 						}
-						sellOrder := reflect.ValueOf(sell).Elem()
-						orderId := sellOrder.FieldByName("OrderId").Int()
-						dash.LogOrder(fmt.Sprintf("[fuchsia::b]TRAILING-STOP MARKET SELL[-] %f %s @ ~[white::b]%.*f[-] %s",
-							qty, scoin, roundPrice, price, dcoin))
-						waitOrderFilled(dash, ticker, orderId, "[fuchsia::b]TRAILING-STOP MARKET SELL[-] filled!", refreshInterval)
+						buyOrder := reflect.ValueOf(buy).Elem()
+						orderId := buyOrder.FieldByName("OrderId").Int()
+						dash.LogOrder(fmt.Sprintf("[fuchsia::b]TRAILING-STOP MARKET BUY[-] %f %s @ ~[white::b]%.*f[-] %s",
+							buyBackQty, scoin, roundPrice, price, dcoin))
+						waitOrderFilled(dash, ticker, orderId, "[fuchsia::b]TRAILING-STOP MARKET BUY[-] filled!", refreshInterval)
 						exitType = "ts"
 						break
 					}
@@ -379,7 +376,7 @@ func bullTradeLoop(
 			// ATR-based dynamic stop-loss: use max(configuredSL, atrMultiplier × ATR%)
 			effectiveSL := stopLoss
 			if cfg.ScalpMode.ATRStopLoss && cfg.Indicators.Atr.Period > 0 {
-				atr := calculateATR(ohlcv.Highs, ohlcv.Lows, ohlcv.Closes, cfg.Indicators.Atr.Period)
+				atr := indicator.CalculateATR(ohlcv.Highs, ohlcv.Lows, ohlcv.Closes, cfg.Indicators.Atr.Period)
 				if len(atr) > 0 {
 					atrMultiplier := cfg.ScalpMode.ATRMultiplier
 					if atrMultiplier <= 0 {
@@ -394,56 +391,58 @@ func bullTradeLoop(
 				}
 			}
 
-			// fixed stop loss (using effective SL which may be ATR-widened)
-			stopLossPrice := buyPrice * (1 - effectiveSL/100)
-			if price <= stopLossPrice {
+			// stop loss: price goes UP (using effective SL which may be ATR-widened)
+			stopLossPrice := sellPrice * (1 + effectiveSL/100)
+			if price >= stopLossPrice {
 				dash.SetPhase("STOP LOSS")
-				sell, err := TradeMarketSell(symbol, roundFloat(qty*0.998, roundAmount), price, roundPrice)
+				buyBackQty := indicator.RoundFloat(sellProceeds/price, roundAmount)
+				buy, err := TradeMarketBuy(symbol, buyBackQty, price, roundPrice)
 				if err != nil {
-					dash.LogError(fmt.Sprintf("Stop-Loss MARKET SELL failed: %v", err))
+					dash.LogError(fmt.Sprintf("Stop-Loss MARKET BUY failed: %v", err))
 					return
 				}
-				sellOrder := reflect.ValueOf(sell).Elem()
-				orderId := sellOrder.FieldByName("OrderId").Int()
-				dash.LogOrder(fmt.Sprintf("[red::b]STOP-LOSS MARKET SELL[-] %f %s @ [white::b]%.*f[-] %s (SL=%.2f%%)",
-					qty, scoin, roundPrice, price, dcoin, effectiveSL))
-				waitOrderFilled(dash, ticker, orderId, "[red::b]STOP-LOSS MARKET SELL[-] filled!", refreshInterval)
+				buyOrder := reflect.ValueOf(buy).Elem()
+				orderId := buyOrder.FieldByName("OrderId").Int()
+				dash.LogOrder(fmt.Sprintf("[red::b]STOP-LOSS MARKET BUY[-] %f %s @ [white::b]%.*f[-] %s (SL=%.2f%%)",
+					buyBackQty, scoin, roundPrice, price, dcoin, effectiveSL))
+				waitOrderFilled(dash, ticker, orderId, "[red::b]STOP-LOSS MARKET BUY[-] filled!", refreshInterval)
 				exitType = "sl"
 				break
 			}
 
 			// take profit with AI exit confirmation
-			profitPrice := buyPrice * (1 + takeProfit/100)
-			var aiSellApproved = true
-			if price >= profitPrice && aiOrch != nil {
+			profitPrice := sellPrice * (1 - takeProfit/100)
+			var aiBuyApproved = true
+			if price <= profitPrice && aiOrch != nil {
 				snapshot := &ai.TechnicalSnapshot{
 					Symbol: symbol, Price: price, PrevPrice: prevPrice,
-					RSI: rsi[len(rsi)-1], Tendency: "sell-exit",
+					RSI: rsi[len(rsi)-1], Tendency: "buy-exit",
 				}
 				ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-				consensus, err := aiOrch.Analyze(ctx, snapshot, "BULL")
+				consensus, err := aiOrch.Analyze(ctx, snapshot, "BEAR")
 				cancel()
 				if err != nil {
-					dash.LogError(fmt.Sprintf("AI sell: %v", err))
+					dash.LogError(fmt.Sprintf("AI buy-back: %v", err))
 				} else {
 					updateDashAI(dash, consensus)
-					aiSellApproved = consensus.ShouldSell() || consensus.FinalSignal == ai.SignalHold
+					aiBuyApproved = consensus.ShouldBuy() || consensus.FinalSignal == ai.SignalHold
 				}
 			}
-			rsiDeclining := rsi[len(rsi)-1] < rsi[len(rsi)-2]
-			rsiExitOk := rsiDeclining || (cfg.ScalpMode.Enabled && !cfg.ScalpMode.RequireRSIExit)
-			if price >= profitPrice && rsiExitOk && aiSellApproved {
+			rsiRising := rsi[len(rsi)-1] > rsi[len(rsi)-2]
+			rsiExitOk := rsiRising || (cfg.ScalpMode.Enabled && !cfg.ScalpMode.RequireRSIExit)
+			if price <= profitPrice && rsiExitOk && aiBuyApproved {
 				dash.SetPhase("TAKE PROFIT")
-				sell, err := TradeSell(symbol, roundFloat(qty*0.998, roundAmount), price, sellFactor, roundPrice)
+				buyBackQty := indicator.RoundFloat(sellProceeds/price, roundAmount)
+				buy, err := TradeBuy(symbol, buyBackQty, price, buyFactor, roundPrice)
 				if err != nil {
-					dash.LogError(fmt.Sprintf("SELL order failed: %v", err))
+					dash.LogError(fmt.Sprintf("BUY order failed: %v", err))
 					return
 				}
-				sellOrder := reflect.ValueOf(sell).Elem()
-				orderId := sellOrder.FieldByName("OrderId").Int()
-				dash.LogOrder(fmt.Sprintf("[red::b]SELL[-] %f %s @ [white::b]%.*f[-] %s = %.*f %s",
-					qty, scoin, roundPrice, price, dcoin, roundPrice, price*qty, dcoin))
-				waitOrderFilled(dash, ticker, orderId, "[red::b]SELL[-] order filled!", refreshInterval)
+				buyOrder := reflect.ValueOf(buy).Elem()
+				orderId := buyOrder.FieldByName("OrderId").Int()
+				dash.LogOrder(fmt.Sprintf("[green::b]BUY[-] %f %s @ [white::b]%.*f[-] %s = %.*f %s",
+					buyBackQty, scoin, roundPrice, price, dcoin, roundPrice, price*buyBackQty, dcoin))
+				waitOrderFilled(dash, ticker, orderId, "[green::b]BUY[-] order filled!", refreshInterval)
 				exitType = "tp"
 				break
 			}
@@ -481,44 +480,5 @@ func bullTradeLoop(
 		}
 		dash.LogInfo(fmt.Sprintf("Operation #%d complete. Next in %ds...", operation-1, interOpDelay))
 		time.Sleep(time.Duration(interOpDelay) * time.Second)
-	}
-}
-
-// updateDashAI converts an ai.ConsensusResult into tui.AIConsensusData and updates the dashboard.
-func updateDashAI(dash *tui.Dashboard, cr *ai.ConsensusResult) {
-	data := &tui.AIConsensusData{
-		FinalSignal:    string(cr.FinalSignal),
-		AvgConfidence:  cr.AvgConfidence,
-		BuyScore:       cr.BuyScore,
-		SellScore:      cr.SellScore,
-		HoldScore:      cr.HoldScore,
-		FearGreed:      -1,
-		FearGreedLabel: "",
-	}
-	if cr.SentimentData != nil {
-		data.FearGreed = cr.SentimentData.FearGreedIndex
-		data.FearGreedLabel = cr.SentimentData.FearGreedLabel
-	}
-	for _, d := range cr.Decisions {
-		data.Agents = append(data.Agents, tui.AgentResult{
-			Provider:   string(d.Provider),
-			Signal:     string(d.Signal),
-			Confidence: d.Confidence,
-			Reasoning:  d.Reasoning,
-		})
-	}
-	dash.UpdateAI(data)
-}
-
-// waitOrderFilled polls until an order is filled, logging the result.
-func waitOrderFilled(dash *tui.Dashboard, ticker string, orderId int64, filledMsg string, interval time.Duration) {
-	for {
-		if getor, err := GetOrder(ticker, orderId); err == nil {
-			if getor.Status == "FILLED" {
-				dash.LogOrder(filledMsg)
-				return
-			}
-		}
-		time.Sleep(interval)
 	}
 }
